+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.18.0 - 2024-02-13
 
 - Remove the translation toolbar button
+- Move the proofreading button to the preview/check-in group
 
 ## 1.17.2 - 2024-02-11